+++ EZOCore.lua
@@ -2,8 +2,8 @@ EZOCore = EZOCore or {}
 local EZOCore = EZOCore
 
 EZOCore.name = "EZOCore"
-EZOCore.version = "0.1.12"
-EZOCore.addOnVersion = 112
+EZOCore.version = "0.1.13"
+EZOCore.addOnVersion = 113
 EZOCore.apiVersion = 1
 
 EZOCore.EVENT_INITIALIZED = "EZO_CORE_INITIALIZED"